+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Please see [CONTRIBUTING.md](./CONTRIBUTING.md) on how to contribute to Cucumber.
 
 ## [Unreleased]
+
+## [12.8.3] - 2026-05-09
 ### Fixed
 - Handle thrown strings correctly ([#2708](https://github.com/cucumber/cucumber-js/pull/2708))
 - Include assertion info in stack trace ([#2803](https://github.com/cucumber/cucumber-js/pull/2803))
@@ -1742,7 +1744,8 @@ this.Given(), this.When(), this.Then() and this.defineStep() ([#2](https://githu
 
 ## 0.0.1
 
-[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v12.8.2...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v12.8.3...HEAD
+[12.8.3]: https://github.com/cucumber/cucumber-js/compare/v12.8.2...v12.8.3
 [12.8.2]: https://github.com/cucumber/cucumber-js/compare/v12.8.1...v12.8.2
 [12.8.1]: https://github.com/cucumber/cucumber-js/compare/v12.8.0...v12.8.1
 [12.8.0]: https://github.com/cucumber/cucumber-js/compare/v12.7.0...v12.8.0